Update McDonalds_data.ipynb
At the begging, where the data is being read using 'pd.read_csv' I changed the path of the .csv file just to its name.
That is because your pathfile file may differ from other's, and if I copy the repo to my local machine and try to run the script it may rise an error. It is better to just put the name of the file in the path and keep the .csv file in one folder with the Jupiter notebook or .py script.